requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-allauth==65.3.0
django-appconf==1.0.6
django-bootstrap5==24.3
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-etc==1.4.0
django-hitcount==1.3.5
django-imagekit==5.0.0
django-js-asset==2.2.0
django-taggit==6.1.0
djangorestframework==3.15.2
Markdown==3.7
pilkit==3.0
pillow==11.0.0
sqlparse==0.5.2
